pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "litecoder"
version = "0.1.0"
description = "Minimal AI coding agent (~1400 LoC) inspired by Claude Code. Works with any LLM. (formerly NanoCoder)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Yufeng He", email = "40085740+he-yufeng@users.noreply.github.com" },
]
keywords = ["ai", "agent", "coding", "cli", "llm", "claude-code"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development",
]
dependencies = [
    "openai>=1.0",
    "rich>=13.0",
    "prompt_toolkit>=3.0",
    "python-dotenv>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["litecoder"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
litecoder = "litecoder.cli:main"


[project.urls]
Homepage = "https://github.com/he-yufeng/CoreCoder"
Repository = "https://github.com/he-yufeng/CoreCoder"
Issues = "https://github.com/he-yufeng/CoreCoder/issues"
